Enable -Wnarrowing in the basic test suite
- Use `size_t` instead of `int` for images width, height and depth
- Update printf strings to use `%zu` instead of `%d` where applicable
- Explicitly cast values to `cl_int` when instantiating an array of values

+++ test_conformance/basic/test_hostptr.cpp
@@ -100,8 +100,8 @@ REGISTER_TEST(hostptr)
     cl_image_format     img_format;
     cl_uchar            *rgba8_inptr, *rgba8_outptr;
     void                *lock_buffer;
-    int                 img_width = 512;
-    int                 img_height = 512;
+    size_t                 img_width = 512;
+    size_t                 img_height = 512;
     cl_int              err;
     MTdata              d;
     RoundingMode        oldRoundMode;
